+++ nym-vpn-core/crates/nym-vpn-lib/src/platform/environment.rs
@@ -1,7 +1,7 @@
 // Copyright 2024 - Nym Technologies SA <contact@nymtech.net>
 // SPDX-License-Identifier: GPL-3.0-only
 
-use crate::uniffi_custom_impls::{AccountLinks, NetworkEnvironment, SystemMessage};
+use super::uniffi_custom_impls::{AccountLinks, NetworkEnvironment, SystemMessage};
 
 use super::{error::VpnError, NETWORK_ENVIRONMENT};
